feat: complete data display components refactoring
Move chart component to data-display/chart directory:
- Created chart component files with proper structure
- Added comprehensive documentation in README.md
- Updated data-display directory README.md and index.ts
- Added backward compatibility exports
- Updated COMPONENT-MOVEMENT-PLAN.md

Now all data display components are properly organized in their own directories.
